[WIP] CMAP SubTable Format4 Bug Fixes
While reading format4 sub table in cmap encoding records, earlier code
was reading `f4->seg_count * 2` bytes for `glyph_id_array` which is an
arbitrary sized array acc. to the spec.

For this, we need to subtract the size of all other fields from the
length value given in format 4 table and then divide by two (size of u16).

This code applies that change.

TODO: reading of variation selector and other fields still remain.
I still need a font with format 2 to understand how that works, finding
sample files for format2 is difficult since that's deprecated.