[WIP] Opaque `XfStructFieldDesc` & `XfStructDesc`
The user-code does not really need to see the internal working or
internal contents of struct and field descriptors. It does not even need
to know there's a field descriptor. We're just exposing structs.

This commit makes these two structs opaque and hard to manipulate
directly. This probably makes the code easier to understand, otherwise
anything that's not opaque hints that it is supposed to be manipulated
somewhere, or it can be alteast.