[WIP] Designing The CrossFileType VM
The earlier solution for implementing a generic binary file format
loader was to implement a langauge to describe the file format and then
compile or transpile it to be used with Xf.

Now the new solution is to implement a VM to which embedded struct
descriptions will be compiled. These embedded descriptions are generated
using macros in the C source code and require almost no effort. They
work by encapsulating the type definition with a macro that defines the
type and stringifies the description at the same time.

The stringified description is then compiled into XftVm bytecode which
can be executed to load the file format. Compiling possibly will be
resource intensive but later on we can serialize and de-serialize the
compiled bytecode to a separate file.

In future, we can also make an attempt to natively compile this and
serialize it as a library file or JIT it at runtime.

For now this new solution looks good because it's not adding an extra
build step to the build system and everything is described and defined
at once place. The descriptions are also quite similar to C (they are
just typedefs with macros).

All these descriptions are also annotated to exetend the descriptions
with some features. These annotations are generated with help of empty
macros. Meaning they're just there for syntactic sugar so that
XftCompiler can understand the description better