[WIP] Grammar Works Without Errors
The fixed TS grammar works without any errors.
Next we write some tests to make sure parsing is as expected
and to interact more with the parse tree.

This knowledge will later on help when dealing with the parse tree in
code.

This is good news because this means parsing part is done. Now either I
created a new VM or hack an existing one with permissive license.